rename chip package to system
