+++ src/main/scala/system/ExampleRocketSystem.scala
@@ -1,6 +1,6 @@
 // See LICENSE.SiFive for license details.
 
-package freechips.rocketchip.chip
+package freechips.rocketchip.system
 
 import Chisel._
 import freechips.rocketchip.config.Parameters
@@ -8,17 +8,17 @@ import freechips.rocketchip.coreplex._
 import freechips.rocketchip.devices.tilelink._
 
 /** Example Top with periphery devices and ports, and a Rocket coreplex */
-class ExampleRocketTop(implicit p: Parameters) extends RocketCoreplex
+class ExampleRocketSystem(implicit p: Parameters) extends RocketCoreplex
     with HasAsyncExtInterrupts
     with HasMasterAXI4MemPort
     with HasMasterAXI4MMIOPort
     with HasSlaveAXI4Port
     with HasPeripheryBootROM
     with HasPeripheryErrorSlave {
-  override lazy val module = new ExampleRocketTopModule(this)
+  override lazy val module = new ExampleRocketSystemModule(this)
 }
 
-class ExampleRocketTopModule[+L <: ExampleRocketTop](_outer: L) extends RocketCoreplexModule(_outer)
+class ExampleRocketSystemModule[+L <: ExampleRocketSystem](_outer: L) extends RocketCoreplexModule(_outer)
     with HasExtInterruptsModuleImp
     with HasMasterAXI4MemPortModuleImp
     with HasMasterAXI4MMIOPortModuleImp
