+++ src/main/scala/system/RocketTestSuite.scala
@@ -1,7 +1,7 @@
 // See LICENSE.Berkeley for license details.
 // See LICENSE.SiFive for license details.
 
-package freechips.rocketchip.chip
+package freechips.rocketchip.system
 
 import scala.collection.mutable.LinkedHashSet
