+++ src/main/scala/system/Utils.scala
@@ -1,6 +1,6 @@
 // See LICENSE.SiFive for license details.
 
-package freechips.rocketchip.chip
+package freechips.rocketchip.system
 
 import scala.collection.mutable.{ArrayBuffer, ListBuffer}
